igraph/requirements.txt
colorcet
matplotlib
tqdm
python-dateutil
requests
igraph